requirements.txt
firebase-admin=6.1.0
binance-connector=3.1.0
pyts=0.12.0
fastbook=0.0.29
schedule=1.2.0